Refactored smart charging tests: * seperated smart_charging_test_utils into hpp and cpp * moved redundant definitions of test cpp files into smart_charging_test_utils * implemented a couple operator== overloads for v21 types
